+++ slipp/src/main/java/slipp/ManualHandlerMapping.java
@@ -2,12 +2,10 @@ package slipp;
 
 import nextstep.mvc.asis.Controller;
 import nextstep.mvc.asis.ControllerAdapter;
-import nextstep.mvc.asis.ForwardController;
 import nextstep.mvc.tobe.handlermapping.HandlerMapping;
 import nextstep.mvc.tobe.handlermapping.annotationmapping.HandlerExecution;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import slipp.controller.*;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.HashMap;
@@ -20,14 +18,6 @@ public class ManualHandlerMapping implements HandlerMapping {
 
     @Override
     public void initialize() {
-        mappings.put("/users/form", new ForwardController("/user/form.jsp"));
-        mappings.put("/users/loginForm", new ForwardController("/user/login.jsp"));
-        mappings.put("/users/profile", new ProfileController());
-        mappings.put("/users/logout", new LogoutController());
-        mappings.put("/users/create", new CreateUserController());
-        mappings.put("/users/updateForm", new UpdateFormUserController());
-        mappings.put("/users/update", new UpdateUserController());
-
         logger.info("Initialized Request Mapping!");
         mappings.keySet().forEach(path -> {
             logger.info("Path : {}, Controller : {}", path, mappings.get(path).getClass());
