--- slipp/src/main/java/slipp/controller/CreateUserController.java
@@ -1,24 +1,0 @@
-package slipp.controller;
-
-import nextstep.mvc.asis.Controller;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import slipp.domain.User;
-import slipp.support.db.DataBase;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-public class CreateUserController implements Controller {
-    private static final Logger log = LoggerFactory.getLogger(CreateUserController.class);
-
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        User user = new User(req.getParameter("userId"), req.getParameter("password"), req.getParameter("name"),
-                req.getParameter("email"));
-        log.debug("User : {}", user);
-
-        DataBase.addUser(user);
-        return "redirect:/";
-    }
-}
