--- slipp/src/main/java/slipp/controller/ListUserController.java
@@ -1,26 +1,0 @@
-package slipp.controller;
-
-import nextstep.mvc.tobe.view.JspView;
-import nextstep.mvc.tobe.view.ModelAndView;
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.annotation.RequestMethod;
-import slipp.support.db.DataBase;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@Controller
-public class ListUserController {
-
-    @RequestMapping(value = "/users", method = RequestMethod.GET)
-    public ModelAndView execute(HttpServletRequest req, HttpServletResponse resp) {
-        if (!UserSessionUtils.isLogined(req.getSession())) {
-            return new ModelAndView(new JspView("redirect:/users/loginForm"));
-        }
-
-        ModelAndView modelAndView = new ModelAndView(new JspView("/user/list.jsp"));
-        modelAndView.addObject("users", DataBase.findAll());
-        return modelAndView;
-    }
-}
