--- slipp/src/main/java/slipp/controller/ProfileController.java
@@ -1,21 +1,0 @@
-package slipp.controller;
-
-import nextstep.mvc.asis.Controller;
-import slipp.domain.User;
-import slipp.support.db.DataBase;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-public class ProfileController implements Controller {
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        String userId = req.getParameter("userId");
-        User user = DataBase.findUserById(userId);
-        if (user == null) {
-            throw new NullPointerException("사용자를 찾을 수 없습니다.");
-        }
-        req.setAttribute("user", user);
-        return "/user/profile.jsp";
-    }
-}
